+++ app/src/main/java/com/pict/pbl/medswift/screens/profile/Profile.kt
@@ -6,7 +6,6 @@ import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Call
 import androidx.compose.material.icons.filled.Person
 import androidx.compose.material.icons.filled.WaterDrop
 import androidx.compose.material3.Icon
@@ -22,11 +21,10 @@ import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.pict.pbl.medswift.R
-import com.pict.pbl.medswift.api.CurrentUser
+import com.pict.pbl.medswift.auth.CurrentUser
 import com.pict.pbl.medswift.ui.theme.MedSwiftTheme
 
 private val currentUser = CurrentUser().getUser()