+++ app/src/main/java/com/pict/pbl/medswift/screens/symptoms/SymptomsActivity.kt
@@ -5,14 +5,28 @@ import android.util.Log
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.viewModels
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Button
+import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.livedata.observeAsState
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.window.Dialog
+import androidx.compose.ui.window.DialogProperties
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
@@ -23,6 +37,9 @@ import com.pict.pbl.medswift.data.Symptom
 import com.pict.pbl.medswift.symptoms.SymptomsJSONReader
 import com.pict.pbl.medswift.ui.theme.MedSwiftTheme
 import com.pict.pbl.medswift.viewmodels.SymptomsViewModel
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.runBlocking
 import okhttp3.OkHttpClient
 
 class SymptomsActivity : ComponentActivity() {
@@ -73,6 +90,7 @@ class SymptomsActivity : ComponentActivity() {
             }
             Button(onClick = {
                 val client = OkHttpClient()
+                symptomsViewModel.isLoading.value = true
                 DiagnosisAPI(client).apply {
                     val analyzeSymptomsList = symptomsViewModel.analyzeSymptomsList.value ?: ArrayList()
                     val mappedSymptoms = HashMap<String,AnalyzeSymptom>()
@@ -85,6 +103,28 @@ class SymptomsActivity : ComponentActivity() {
                 }
             }) {
                 Text(text = "Analyze")
+            }
+            ProgressDialog()
+        }
+    }
+
+    @Composable
+    private fun ProgressDialog() {
+        val progressShowFlag by symptomsViewModel.isLoading.observeAsState()
+        var openDialog by rememberSaveable{ mutableStateOf( true ) }
+        if (openDialog && progressShowFlag == true ) {
+            Dialog(
+                onDismissRequest = { openDialog = false },
+                DialogProperties(dismissOnBackPress = false, dismissOnClickOutside = false)
+            ) {
+                Box(
+                    contentAlignment= Alignment.Center,
+                    modifier = Modifier
+                        .size(100.dp)
+                        .background(Color.White, shape = RoundedCornerShape(8.dp))
+                ) {
+                    CircularProgressIndicator()
+                }
             }
         }
     }